+++ core/parser/src/main/kotlin/fr/forumhfr/redface2/core/parser/write/ReplyFormParser.kt
@@ -2,6 +2,7 @@ package fr.forumhfr.redface2.core.parser.write
 
 import fr.forumhfr.redface2.core.model.write.ReplyForm
 import fr.forumhfr.redface2.core.model.write.ReplyFormOptions
+import fr.forumhfr.redface2.core.parser.smiley.SmileyUserIdExtractor
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Element
 
@@ -79,6 +80,7 @@ class ReplyFormParser {
                 initialContent = initialContent,
                 options = parseOptions(replyForm),
                 msgIcon = parseMsgIcon(replyForm),
+                userId = SmileyUserIdExtractor.extract(html),
             ),
         )
     }