+++ core/parser/src/main/kotlin/fr/forumhfr/redface2/core/parser/write/TopicFormParser.kt
@@ -4,6 +4,7 @@ import fr.forumhfr.redface2.core.model.write.ReplyFormOptions
 import fr.forumhfr.redface2.core.model.write.TopicForm
 import fr.forumhfr.redface2.core.model.write.TopicFormSubcategoryChoice
 import fr.forumhfr.redface2.core.model.write.TopicPollForm
+import fr.forumhfr.redface2.core.parser.smiley.SmileyUserIdExtractor
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Element
 
@@ -101,6 +102,7 @@ class TopicFormParser {
                 hashCheck = hashCheck,
                 subject = subject,
                 initialContent = initialContent,
+                userId = SmileyUserIdExtractor.extract(html),
                 selectedSubcat = selectedSubcat,
                 subcategoryChoices = subcategoryChoices,
                 hiddenFields = collection.fields,